Update draft-rokui-ccamp-actn-wdm-pluggable-modelling.md
Adressing Terminology of the main textup to the appendix which needs to be checked separately.:
- If the context is about pluggability and/or physical form factor, call it -> pluggable module
- if the context is about optics, call it ->  optical module
- If the context is strictly about coherent, call it --> coherent optical module
In general, the term "pluggable" has been changed to "pluggable module" or "optical module" or "coherent optical module" depending on the context.
Note that  large number of optical parameters is not related to coherent technology. Likewise, many optical modules are controlled through CMIS irrespective of their technology or being pluggable.